Restructure of components and main app
